requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# ML & Embeddings
sentence-transformers==2.2.2
faiss-cpu==1.8.0
rank-bm25==0.2.2
numpy>=1.26.0
scikit-learn==1.3.2

# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# LLM Integration
groq==0.4.1
google-generativeai==0.3.1

# Data Processing
pandas==2.1.3
openpyxl==3.1.2

# Utilities
tqdm==4.66.1
aiofiles==23.2.1
